timetable-kafka/requirements.txt
kafka-python
requests
xmltodict
tzdata